+++ e2e/cluster/cluster.go
@@ -5,26 +5,12 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	"net"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/Azure/agentbakere2e/config"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/network/armnetwork"
-	appsv1 "k8s.io/api/apps/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/yaml"
 )
 
 // WARNING: if you modify cluster configuration, please change the version below
@@ -53,12 +39,6 @@ type Cluster struct {
 	SubnetID string
 }
 
-type Kubeclient struct {
-	Dynamic client.Client
-	Typed   kubernetes.Interface
-	Rest    *rest.Config
-}
-
 // Returns true if the cluster is configured with Azure CNI
 func (c *Cluster) IsAzureCNI() (bool, error) {
 	if c.Model.Properties.NetworkProfile != nil {
@@ -79,14 +59,14 @@ func (c *Cluster) MaxPodsPerNode() (int, error) {
 // sync.Once is used to ensure that only one cluster for the set of tests is created
 func ClusterKubenet(ctx context.Context) (*Cluster, error) {
 	clusterKubenetOnce.Do(func() {
-		clusterKubenet, clusterKubenetError = createNewClusterWithClient(ctx, getKubenetClusterModel(testClusterNamePrefix+"kubenet-v1"))
+		clusterKubenet, clusterKubenetError = createCluster(ctx, getKubenetClusterModel(testClusterNamePrefix+"kubenet-v1"))
 	})
 	return clusterKubenet, clusterKubenetError
 }
 
 func ClusterKubenetAirgap(ctx context.Context) (*Cluster, error) {
 	clusterKubenetAirgapOnce.Do(func() {
-		cluster, err := createNewClusterWithClient(ctx, getKubenetClusterModel(testClusterNamePrefix+"kubenet-airgap"))
+		cluster, err := createCluster(ctx, getKubenetClusterModel(testClusterNamePrefix+"kubenet-airgap"))
 		if err == nil {
 			err = addAirgapNetworkSettings(ctx, cluster)
 		}
@@ -97,13 +77,13 @@ func ClusterKubenetAirgap(ctx context.Context) (*Cluster, error) {
 
 func ClusterAzureNetwork(ctx context.Context) (*Cluster, error) {
 	clusterAzureNetworkOnce.Do(func() {
-		clusterAzureNetwork, clusterAzureNetworkError = createNewClusterWithClient(ctx, getAzureNetworkClusterModel(testClusterNamePrefix+"azure-network"))
+		clusterAzureNetwork, clusterAzureNetworkError = createCluster(ctx, getAzureNetworkClusterModel(testClusterNamePrefix+"azure-network"))
 	})
 	return clusterAzureNetwork, clusterAzureNetworkError
 }
 
-func createNewClusterWithClient(ctx context.Context, cluster *armcontainerservice.ManagedCluster) (*Cluster, error) {
-	createdCluster, err := createNewClusterWithRetry(ctx, cluster)
+func createCluster(ctx context.Context, cluster *armcontainerservice.ManagedCluster) (*Cluster, error) {
+	createdCluster, err := createNewAKSClusterWithRetry(ctx, cluster)
 	if err != nil {
 		return nil, err
 	}
@@ -126,7 +106,7 @@ func createNewClusterWithClient(ctx context.Context, cluster *armcontainerservic
 
 }
 
-func createNewCluster(ctx context.Context, cluster *armcontainerservice.ManagedCluster) (*armcontainerservice.ManagedCluster, error) {
+func createNewAKSCluster(ctx context.Context, cluster *armcontainerservice.ManagedCluster) (*armcontainerservice.ManagedCluster, error) {
 	log.Printf("Creating new cluster %s in rg %s\n", *cluster.Name, *cluster.Location)
 	pollerResp, err := config.Azure.AKS.BeginCreateOrUpdate(
 		ctx,
@@ -147,18 +127,18 @@ func createNewCluster(ctx context.Context, cluster *armcontainerservice.ManagedC
 	return &clusterResp.ManagedCluster, nil
 }
 
-// createNewClusterWithRetry is a wrapper around createNewCluster
+// createNewAKSClusterWithRetry is a wrapper around createNewAKSCluster
 // that retries creating a cluster if it fails with a 409 Conflict error
 // clusters are reused, and sometimes a cluster can be in UPDATING or DELETING state
 // simple retry should be sufficient to avoid such conflicts
-func createNewClusterWithRetry(ctx context.Context, cluster *armcontainerservice.ManagedCluster) (*armcontainerservice.ManagedCluster, error) {
+func createNewAKSClusterWithRetry(ctx context.Context, cluster *armcontainerservice.ManagedCluster) (*armcontainerservice.ManagedCluster, error) {
 	maxRetries := 10
 	retryInterval := 30 * time.Second
 	var lastErr error
 	for attempt := 0; attempt < maxRetries; attempt++ {
 		log.Printf("Attempt %d: Creating new cluster %s in rg %s\n", attempt+1, *cluster.Name, *cluster.Location)
 
-		createdCluster, err := createNewCluster(ctx, cluster)
+		createdCluster, err := createNewAKSCluster(ctx, cluster)
 		if err == nil {
 			return createdCluster, nil
 		}
@@ -184,87 +164,6 @@ func createNewClusterWithRetry(ctx context.Context, cluster *armcontainerservice
 	return nil, fmt.Errorf("failed to create cluster after %d attempts due to persistent 409 Conflict: %w", maxRetries, lastErr)
 }
 
-func getKubenetClusterModel(name string) *armcontainerservice.ManagedCluster {
-	model := getBaseClusterModel(name)
-	model.Properties.NetworkProfile.NetworkPlugin = to.Ptr(armcontainerservice.NetworkPluginKubenet)
-	return model
-}
-
-func getAzureNetworkClusterModel(name string) *armcontainerservice.ManagedCluster {
-	cluster := getBaseClusterModel(name)
-	cluster.Properties.NetworkProfile.NetworkPlugin = to.Ptr(armcontainerservice.NetworkPluginAzure)
-	if cluster.Properties.AgentPoolProfiles != nil {
-		for _, app := range cluster.Properties.AgentPoolProfiles {
-			app.MaxPods = to.Ptr[int32](30)
-		}
-	}
-	return cluster
-}
-
-func getBaseClusterModel(clusterName string) *armcontainerservice.ManagedCluster {
-	return &armcontainerservice.ManagedCluster{
-		Name:     to.Ptr(clusterName),
-		Location: to.Ptr(config.Location),
-		Properties: &armcontainerservice.ManagedClusterProperties{
-			DNSPrefix: to.Ptr(clusterName),
-			AgentPoolProfiles: []*armcontainerservice.ManagedClusterAgentPoolProfile{
-				{
-					Name:         to.Ptr("nodepool1"),
-					Count:        to.Ptr[int32](2),
-					VMSize:       to.Ptr("standard_d2s_v4"),
-					MaxPods:      to.Ptr[int32](110),
-					OSType:       to.Ptr(armcontainerservice.OSTypeLinux),
-					Type:         to.Ptr(armcontainerservice.AgentPoolTypeVirtualMachineScaleSets),
-					Mode:         to.Ptr(armcontainerservice.AgentPoolModeSystem),
-					OSDiskSizeGB: to.Ptr[int32](512),
-				},
-			},
-			NetworkProfile: &armcontainerservice.NetworkProfile{
-				NetworkPlugin: to.Ptr(armcontainerservice.NetworkPluginKubenet),
-			},
-		},
-		Identity: &armcontainerservice.ManagedClusterIdentity{
-			Type: to.Ptr(armcontainerservice.ResourceIdentityTypeSystemAssigned),
-		},
-	}
-}
-
-func addAirgapNetworkSettings(ctx context.Context, cluster *Cluster) error {
-	log.Printf("Adding network settings for airgap cluster %s in rg %s\n", *cluster.Model.Name, *cluster.Model.Properties.NodeResourceGroup)
-
-	vnet, err := getClusterVNet(ctx, *cluster.Model.Properties.NodeResourceGroup)
-	if err != nil {
-		return err
-	}
-	subnetId := vnet.subnetId
-
-	nsgParams, err := airGapSecurityGroup(config.Location, *cluster.Model.Properties.Fqdn)
-	if err != nil {
-		return err
-	}
-
-	nsg, err := createAirgapSecurityGroup(ctx, cluster.Model, nsgParams, nil)
-	if err != nil {
-		return err
-	}
-
-	subnetParameters := armnetwork.Subnet{
-		ID: to.Ptr(subnetId),
-		Properties: &armnetwork.SubnetPropertiesFormat{
-			AddressPrefix: to.Ptr("10.224.0.0/16"),
-			NetworkSecurityGroup: &armnetwork.SecurityGroup{
-				ID: nsg.ID,
-			},
-		},
-	}
-	if err = updateSubnet(ctx, cluster.Model, subnetParameters, vnet.name); err != nil {
-		return err
-	}
-
-	log.Printf("updated cluster %s subnet with airggap settings", *cluster.Model.Name)
-	return nil
-}
-
 type VNet struct {
 	name     string
 	subnetId string
@@ -285,242 +184,4 @@ func getClusterVNet(ctx context.Context, mcResourceGroupName string) (VNet, erro
 		}
 	}
 	return VNet{}, fmt.Errorf("failed to find aks vnet")
-}
-
-func airGapSecurityGroup(location, clusterFQDN string) (armnetwork.SecurityGroup, error) {
-	requiredRules, err := getRequiredSecurityRules(clusterFQDN)
-	if err != nil {
-		return armnetwork.SecurityGroup{}, fmt.Errorf("failed to get required security rules for airgap resource group: %w", err)
-	}
-
-	allowVnet := &armnetwork.SecurityRule{
-		Name: to.Ptr("AllowVnetOutBound"),
-		Properties: &armnetwork.SecurityRulePropertiesFormat{
-			Protocol:                 to.Ptr(armnetwork.SecurityRuleProtocolAsterisk),
-			Access:                   to.Ptr(armnetwork.SecurityRuleAccessAllow),
-			Direction:                to.Ptr(armnetwork.SecurityRuleDirectionOutbound),
-			SourceAddressPrefix:      to.Ptr("VirtualNetwork"),
-			SourcePortRange:          to.Ptr("*"),
-			DestinationAddressPrefix: to.Ptr("VirtualNetwork"),
-			DestinationPortRange:     to.Ptr("*"),
-			Priority:                 to.Ptr[int32](2000),
-		},
-	}
-
-	blockOutbound := &armnetwork.SecurityRule{
-		Name: to.Ptr("block-all-outbound"),
-		Properties: &armnetwork.SecurityRulePropertiesFormat{
-			Protocol:                 to.Ptr(armnetwork.SecurityRuleProtocolAsterisk),
-			Access:                   to.Ptr(armnetwork.SecurityRuleAccessDeny),
-			Direction:                to.Ptr(armnetwork.SecurityRuleDirectionOutbound),
-			SourceAddressPrefix:      to.Ptr("*"),
-			SourcePortRange:          to.Ptr("*"),
-			DestinationAddressPrefix: to.Ptr("*"),
-			DestinationPortRange:     to.Ptr("*"),
-			Priority:                 to.Ptr[int32](2001),
-		},
-	}
-
-	rules := append([]*armnetwork.SecurityRule{allowVnet, blockOutbound}, requiredRules...)
-
-	return armnetwork.SecurityGroup{
-		Location:   &location,
-		Name:       &config.AirgapNSGName,
-		Properties: &armnetwork.SecurityGroupPropertiesFormat{SecurityRules: rules},
-	}, nil
-}
-
-func getRequiredSecurityRules(clusterFQDN string) ([]*armnetwork.SecurityRule, error) {
-	// https://learn.microsoft.com/en-us/azure/aks/outbound-rules-control-egress#azure-global-required-fqdn--application-rules
-	// note that we explicitly exclude packages.microsoft.com
-	requiredDNSNames := []string{
-		"mcr.microsoft.com",
-		"management.azure.com",
-		"acs-mirror.azureedge.net",
-		clusterFQDN,
-	}
-	var rules []*armnetwork.SecurityRule
-	var priority int32 = 100
-
-	for _, dnsName := range requiredDNSNames {
-		ips, err := net.LookupIP(dnsName)
-		if err != nil {
-			return nil, fmt.Errorf("failed to lookup IP for DNS name %s: %w", dnsName, err)
-		}
-		for _, ip := range ips {
-			if ipv4 := ip.To4(); ipv4 != nil {
-				rules = append(rules, getSecurityRule(fmt.Sprintf("%s-%d", strings.ReplaceAll(dnsName, ".", "-"), priority), ipv4.String(), priority))
-				priority++
-			}
-		}
-	}
-
-	return rules, nil
-}
-
-func getSecurityRule(name, destinationAddressPrefix string, priority int32) *armnetwork.SecurityRule {
-	return &armnetwork.SecurityRule{
-		Name: to.Ptr(name),
-		Properties: &armnetwork.SecurityRulePropertiesFormat{
-			Protocol:                 to.Ptr(armnetwork.SecurityRuleProtocolAsterisk),
-			Access:                   to.Ptr(armnetwork.SecurityRuleAccessAllow),
-			Direction:                to.Ptr(armnetwork.SecurityRuleDirectionOutbound),
-			SourceAddressPrefix:      to.Ptr("*"),
-			SourcePortRange:          to.Ptr("*"),
-			DestinationAddressPrefix: to.Ptr(destinationAddressPrefix),
-			DestinationPortRange:     to.Ptr("*"),
-			Priority:                 to.Ptr[int32](priority),
-		},
-	}
-}
-
-func createAirgapSecurityGroup(ctx context.Context, cluster *armcontainerservice.ManagedCluster, nsgParams armnetwork.SecurityGroup, options *armnetwork.SecurityGroupsClientBeginCreateOrUpdateOptions) (*armnetwork.SecurityGroupsClientCreateOrUpdateResponse, error) {
-	poller, err := config.Azure.SecurityGroup.BeginCreateOrUpdate(ctx, *cluster.Properties.NodeResourceGroup, config.AirgapNSGName, nsgParams, options)
-	if err != nil {
-		return nil, err
-	}
-	nsg, err := poller.PollUntilDone(ctx, nil)
-	if err != nil {
-		return nil, err
-	}
-	return &nsg, nil
-}
-
-func updateSubnet(ctx context.Context, cluster *armcontainerservice.ManagedCluster, subnetParameters armnetwork.Subnet, vnetName string) error {
-	poller, err := config.Azure.Subnet.BeginCreateOrUpdate(ctx, *cluster.Properties.NodeResourceGroup, vnetName, config.DefaultSubnetName, subnetParameters, nil)
-	if err != nil {
-		return err
-	}
-	_, err = poller.PollUntilDone(ctx, nil)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func newKubeclient(config *rest.Config) (*Kubeclient, error) {
-	dynamic, err := client.New(config, client.Options{})
-	if err != nil {
-		return nil, fmt.Errorf("create dynamic Kubeclient: %w", err)
-	}
-
-	restClient, err := rest.RESTClientFor(config)
-	if err != nil {
-		return nil, fmt.Errorf("create rest kube client: %w", err)
-	}
-
-	typed := kubernetes.New(restClient)
-
-	return &Kubeclient{
-		Dynamic: dynamic,
-		Typed:   typed,
-		Rest:    config,
-	}, nil
-}
-
-func getClusterKubeClient(ctx context.Context, resourceGroupName, clusterName string) (*Kubeclient, error) {
-	data, err := getClusterKubeconfigBytes(ctx, resourceGroupName, clusterName)
-	if err != nil {
-		return nil, fmt.Errorf("get cluster kubeconfig bytes: %w", err)
-	}
-
-	restConfig, err := clientcmd.RESTConfigFromKubeConfig(data)
-	if err != nil {
-		return nil, fmt.Errorf("convert kubeconfig bytes to rest config: %w", err)
-	}
-	restConfig.NegotiatedSerializer = serializer.WithoutConversionCodecFactory{CodecFactory: scheme.Codecs}
-	restConfig.APIPath = "/api"
-	restConfig.GroupVersion = &schema.GroupVersion{
-		Version: "v1",
-	}
-
-	return newKubeclient(restConfig)
-}
-
-func getClusterKubeconfigBytes(ctx context.Context, resourceGroupName, clusterName string) ([]byte, error) {
-	credentialList, err := config.Azure.AKS.ListClusterAdminCredentials(ctx, resourceGroupName, clusterName, nil)
-	if err != nil {
-		return nil, fmt.Errorf("list cluster admin credentials: %w", err)
-	}
-
-	if len(credentialList.Kubeconfigs) < 1 {
-		return nil, fmt.Errorf("no kubeconfigs available for the managed cluster cluster")
-	}
-
-	return credentialList.Kubeconfigs[0].Value, nil
-}
-
-// this is a bit ugly, but we don't want to execute this piece concurrently with other tests
-func ensureDebugDaemonset(ctx context.Context, kube *Kubeclient) error {
-	manifest := getDebugDaemonset()
-	var ds appsv1.DaemonSet
-
-	if err := yaml.Unmarshal([]byte(manifest), &ds); err != nil {
-		return fmt.Errorf("failed to unmarshal debug daemonset manifest: %w", err)
-	}
-
-	desired := ds.DeepCopy()
-	_, err := controllerutil.CreateOrUpdate(ctx, kube.Dynamic, &ds, func() error {
-		ds = *desired
-		return nil
-	})
-
-	if err != nil {
-		return fmt.Errorf("failed to apply debug daemonset: %w", err)
-	}
-
-	return nil
-}
-
-func getDebugDaemonset() string {
-	return `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: &name debug
-  namespace: default
-  labels:
-    app: *name
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: *name
-  template:
-    metadata:
-      labels:
-        app: *name
-    spec:
-      hostNetwork: true
-      nodeSelector:
-        kubernetes.azure.com/agentpool: nodepool1
-      hostPID: true
-      containers:
-      - image: mcr.microsoft.com/oss/nginx/nginx:1.21.6
-        name: ubuntu
-        command: ["sleep", "infinity"]
-        resources:
-          requests: {}
-          limits: {}
-        securityContext:
-          privileged: true
-          capabilities:
-            add: ["SYS_PTRACE", "SYS_RAWIO"]
-`
-}
-
-func getClusterSubnetID(ctx context.Context, mcResourceGroupName string) (string, error) {
-	pager := config.Azure.VNet.NewListPager(mcResourceGroupName, nil)
-	for pager.More() {
-		nextResult, err := pager.NextPage(ctx)
-		if err != nil {
-			return "", fmt.Errorf("advance page: %w", err)
-		}
-		for _, v := range nextResult.Value {
-			if v == nil {
-				return "", fmt.Errorf("aks vnet was empty")
-			}
-			return fmt.Sprintf("%s/subnets/%s", *v.ID, "aks-subnet"), nil
-		}
-	}
-	return "", fmt.Errorf("failed to find aks vnet")
 }
